+++ dotnet/Apps/Database/Server/Program.cs
@@ -1,7 +1,6 @@
 namespace Allors.Database.Server.Controllers
 {
     using System;
-    using System.IO;
     using Configuration;
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.Extensions.Hosting;
@@ -9,8 +8,6 @@ namespace Allors.Database.Server.Controllers
 
     public class Program
     {
-        private const string ConfigPath = "/config/apps";
-
         public static void Main(string[] args)
         {
             var logger = NLogBuilder.ConfigureNLog("nlog.config").GetCurrentClassLogger();
@@ -37,9 +34,7 @@ namespace Allors.Database.Server.Controllers
                     .ConfigureAppConfiguration((hostingContext, configurationBuilder) =>
                     {
                         var environmentName = hostingContext.HostingEnvironment.EnvironmentName;
-                        configurationBuilder.AddCrossPlatform(".", environmentName, true);
-                        configurationBuilder.AddCrossPlatform(ConfigPath, environmentName);
-                        configurationBuilder.AddCrossPlatform(Path.Combine(ConfigPath, hostingContext.HostingEnvironment.ApplicationName), environmentName);
+                        configurationBuilder.AddAllorsConfiguration("apps", hostingContext.HostingEnvironment.ApplicationName, environmentName);
                     })
                     .UseNLog());
     }