+++ dotnet/System/Database/Adapters/Tests.Static/Static/Sql/SqlClient/Fixture.cs
@@ -1,26 +1,12 @@
-// <copyright file="Profile.cs" company="Allors bv">
+// <copyright file="Fixture.cs" company="Allors bv">
 // Copyright (c) Allors bv. All rights reserved.
 // Licensed under the LGPL license. See LICENSE file in the project root for full license information.
 // </copyright>
 
 namespace Allors.Database.Adapters.Sql.SqlClient
 {
-    using Microsoft.Data.SqlClient;
-
     public class Fixture<T>
     {
-        public Fixture()
-        {
-            var database = typeof(T).Name;
-
-            using var connection = new SqlConnection(@"Server=(localdb)\MSSQLLocalDB;Database=master;Integrated Security=true;Connection Timeout=0");
-            connection.Open();
-            using var command = connection.CreateCommand();
-            command.CommandTimeout = 300;
-            command.CommandText = $"DROP DATABASE IF EXISTS {database}";
-            command.ExecuteNonQuery();
-            command.CommandText = $"CREATE DATABASE {database}";
-            command.ExecuteNonQuery();
-        }
+        public Fixture() => Provisioning.DropCreate(typeof(T).Name);
     }
 }